Fixed Crate Spawning
The crates just needed to be removed from the scene, added to the Network Manager Prefab list and then readded to the scene. The colours of the crates don't quite seem to sync across the network correctly, this could be achieved with a NetworkVariable but needs more testing as it doesn't have a pattern. (Perhaps each client is generating it's own colours for each crate)