Started Creating a Level
Extended the Platform Arranger to create raised platforms which can be various heights, widths and scales. Created a script to easily add more links to a chain. Created the basics of a level with fixed platforms, a giant chain, a swinging platform and a crane with 2 zones. A lot more work needs to be done on this level but we should now have some templates for different things that can feature in levels.