Created Moving Platforms and One Way Platforms
The moving platforms don't work amazingly as they are currently applying a force to the player to keep them stuck to the platform which means that hands and other body parts can get stuck to the platforms. The one way platforms work really well although need a little bit of tweaking and I might even prefer them to the regular platforms.