Jump off other players but not grab them
The grabbing didn't add anything to the game although was a bit of fun. There was a big problem with being able to jump as your own foot is triggered by the overlapcircle but this is fixed by checking that it's parent is not the script calling the function although it now needs to check all colliders.