FINALLY Fixed the Arrow Manager
The arrow manager will now correctly find all of the objects in the scene that match the current objective object and then point the player to the nearest one. This will also be updated if the nearest one changes. Also added a small fix to the ScoreManager to update the UI correctly when increasing the streak